+++ packages/util/index.ts
@@ -1,2 +1,3 @@
 export { getHTMLElement } from "./get-html-element";
 export { assertNotNull, assertNotUndefined, hardFail } from "./assertions";
+export { storageFactory } from "./storage-factory";