+++ packages/util/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.4.0 (2022-12-21)
+
+Add `storageFactory` function to help with avoid
+errors with localStorage and sessionStorage when 
+they aren't supported by the browser settings.
+
 ## 0.3.0 (2021-04-27)
 
 Add an optional custom message to assertion helpers `assertNotUndefined` and `assertNotNull`.